Decompile func_80323164

+++ src/app/code_A9D00.h
@@ -7,7 +7,7 @@ void func_802E682C(f32*, s32, u8);
 void func_802E6870(f32*, s32, u8*);
 void func_80322DA8(s32);
 f32 func_803230B0(s32 contIdx, s32 axis);
-s32 func_80323164(u16);
+s32 func_80323164(s32);
 s32 func_803231A0(s32, s32);
 s32 func_803231E0(u16, s32);
